backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
google-generativeai==0.3.0
python-dotenv==1.0.1
numpy==1.24.3
